Organize pojos in packages based on their output file
