+++ src/main/java/org/dspace/pack/bagit/BagItAipReader.java
@@ -18,15 +18,12 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
 import java.util.Properties;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import javax.xml.stream.XMLInputFactory;
-import javax.xml.stream.XMLStreamConstants;
 import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamReader;
 
@@ -39,12 +36,10 @@ import gov.loc.repository.bagit.exceptions.UnsupportedAlgorithmException;
 import gov.loc.repository.bagit.reader.BagReader;
 import gov.loc.repository.bagit.verify.BagVerifier;
 import org.apache.commons.io.FileUtils;
-import org.dspace.pack.bagit.xml.Element;
-import org.dspace.pack.bagit.xml.Metadata;
-import org.dspace.pack.bagit.xml.MetadataDeserializer;
-import org.dspace.pack.bagit.xml.Policy;
-import org.dspace.pack.bagit.xml.PolicyDeserializer;
-import org.dspace.pack.bagit.xml.Value;
+import org.dspace.pack.bagit.xml.metadata.Metadata;
+import org.dspace.pack.bagit.xml.metadata.MetadataDeserializer;
+import org.dspace.pack.bagit.xml.policy.Policy;
+import org.dspace.pack.bagit.xml.policy.PolicyDeserializer;
 import org.dspace.services.ConfigurationService;
 import org.dspace.services.factory.DSpaceServicesFactory;
 import org.duraspace.bagit.BagDeserializer;
