+++ src/main/java/org/dspace/pack/bagit/BagItAipWriter.java
@@ -42,8 +42,8 @@ import org.dspace.core.Context;
 import org.dspace.core.Utils;
 import org.dspace.curate.Curator;
 import org.dspace.pack.bagit.xml.Element;
-import org.dspace.pack.bagit.xml.Metadata;
-import org.dspace.pack.bagit.xml.Policy;
+import org.dspace.pack.bagit.xml.metadata.Metadata;
+import org.dspace.pack.bagit.xml.policy.Policy;
 import org.dspace.services.ConfigurationService;
 import org.dspace.services.factory.DSpaceServicesFactory;
 import org.duraspace.bagit.BagConfig;
