+++ src/main/java/org/dspace/pack/bagit/BagItPolicyUtil.java
@@ -37,8 +37,7 @@ import org.dspace.eperson.factory.EPersonServiceFactory;
 import org.dspace.eperson.service.EPersonService;
 import org.dspace.eperson.service.GroupService;
 import org.dspace.pack.bagit.xml.Element;
-import org.dspace.pack.bagit.xml.Metadata;
-import org.dspace.pack.bagit.xml.Policy;
+import org.dspace.pack.bagit.xml.policy.Policy;
 import org.dspace.pack.bagit.xml.Value;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
