+++ src/main/java/org/dspace/pack/bagit/ItemPacker.java
@@ -45,8 +45,8 @@ import org.dspace.core.Context;
 import org.dspace.curate.Curator;
 import org.dspace.pack.Packer;
 import org.dspace.pack.bagit.xml.Element;
-import org.dspace.pack.bagit.xml.Metadata;
-import org.dspace.pack.bagit.xml.Policy;
+import org.dspace.pack.bagit.xml.metadata.Metadata;
+import org.dspace.pack.bagit.xml.policy.Policy;
 import org.dspace.pack.bagit.xml.Value;
 
 /**
